crawlingServer/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
schedule==1.2.0
mysql-connector-python==8.0.33
